scripts/pyproject.toml
[project]
name = "qualcomm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "accelerate>=1.7.0",
    "argparse>=1.4.0",
    "diffusers>=0.33.1",
    "fastapi>=0.115.12",
    "huggingface-hub>=0.33.0",
    "numpy>=1.23.5,<2.0.0",
    "opencv-python>=4.5.0",
    "pillow>=8.0.0,<12.0.0",
    "python-multipart>=0.0.20",
    "qai-hub-models[midas]>=0.30.4",
    "rich>=14.0.0",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.0",
    "torch>=1.10.0,<2.5.0",
    "torchvision>=0.16.0,<0.21.0",
    "transformers>=4.20.0",
    "ultralytics>=8.0.0",
    "uvicorn>=0.34.3",
    "vidstab>=1.7.4",
    "yt-dlp>=2025.6.9",
]
